Main components for a quest.
Quest - stores relevant data and tracks it's state.
QuestStage - A stage of quest that contains an array of objectives. Gets completed when all objectives are completed and retrieves the concurrent stage to progress in the quest.
QuestObjective - the goal needed to be achieved. Can be marked optional to be ignored or marked as critical fail to force the quest to close upon failing the objective.